fixed and added fields

+++ src/main/java/com/ToDoList/model/repository/SubtaskRepository.java
@@ -3,5 +3,5 @@ package com.ToDoList.model.repository;
 import com.ToDoList.model.repository.entity.Subtask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubtaskRepository extends JpaRepository<Subtask, Integer> {
+public interface SubtaskRepository extends JpaRepository<Subtask, Long> {
 }
